normalize route options in the router
